requirements.txt
alembic==1.13.1
bcrypt==4.1.2
blinker==1.7.0
cached-property==1.5.2
cachelib==0.9.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Caching==2.1.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
idna==3.7
inflection==0.5.1
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.3
Mako==1.3.3
MarkupSafe==2.1.5
marshmallow==3.21.2
mypy-extensions==1.0.0
packaging==24.0
pluggy==1.5.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytest==8.1.1
python-dotenv==1.0.1
requests==2.31.0
SQLAlchemy==2.0.29
sqlalchemy-orm==1.2.10
typing-inspect==0.9.0
typing_extensions==4.11.0
urllib3==2.2.1
webargs==8.4.0
Werkzeug==3.0.2
